+++ factual/query/__init__.py
@@ -6,7 +6,6 @@ from .insert import Insert
 from .clear import Clear
 from .facets import Facets
 from .flag import Flag
-from .geopulse import Geopulse
 from .geocode import Geocode
 from .diffs import Diffs
 from .multi import Multi